Refactor learn page and add tutorial subpages
Simplifies pages/learn.py to a landing page with links to tutorials.

Moves Quickstart and Tutorial 1 content into new dedicated files: pages/learn_quickstart.py and pages/learn_tutorial_1.py for better maintainability, and to encourage the writing of more tutorials.
